split ts js test application
